Begin work on sensors, make most requested changes
