+++ actions/src/openstack_check_actions.py
@@ -1,5 +1,4 @@
 import json
-import os
 from typing import Callable, Dict
 import logging
 import sys
@@ -8,7 +7,6 @@ import requests
 from openstack_api.openstack_connection import OpenstackConnection
 from st2common.runners.base_action import Action
 from requests.auth import HTTPBasicAuth
-from amphorae import get_amphorae
 
 
 class CheckActions(Action):
@@ -203,143 +201,12 @@ class CheckActions(Action):
 
         return output
 
-    def loadbalancer_check(self, cloud_account: str):
-        """
-        Action to check loadbalancer and amphorae status
-        output suitable to be passed to create_tickets
-        """
-
-        amphorae = get_amphorae(cloud_account)
-
-        amph_json = self._check_amphora_status(amphorae)
-        # pylint: disable=line-too-long
-        output = {
-            "title": "{p[title_text]}",
-            "body": "The loadbalance ping test result was: {p[lb_status]}\nThe status of the amphora was: {p[amp_status]}\nThe amphora id is: {p[amp_id]}\nThe loadbalancer id is: {p[lb_id]}",
-            "server_list": [],
-        }
-        if amphorae.status_code == 403:
-            # Checks user has permissions to access api
-            logging.critical("Please run this script using a cloud admin account.")
-            return "Please run this script using a cloud admin account."
-        if amphorae.status_code == 200:
-            # Gets list of amphorae and iterates through it to check the loadbalancer and amphora status.
-            for i in amph_json["amphorae"]:
-                status = self._check_status(i)
-                ping_result = self._ping_lb(i["lb_network_ip"])
-                # This section builds out the ticket for each one with an error
-                if status[0] == "error" or ping_result == "error":
-                    if status[0].lower() == "error" and ping_result.lower() == "error":
-                        output["server_list"].append(
-                            {
-                                "dataTitle": {
-                                    "title_text": "Issue with loadbalancer "
-                                    + str(i["loadbalancer_id"] or "null")
-                                    + " and amphora "
-                                    + str(i["id"] or "null"),
-                                    "lb_id": str(i["loadbalancer_id"] or "null"),
-                                    "amp_id": str(i["id"] or "null"),
-                                },
-                                "dataBody": {
-                                    "lb_status": str(ping_result or "null"),
-                                    "amp_status": str(status[1] or "null"),
-                                    "lb_id": str(i["loadbalancer_id"] or "null"),
-                                    "amp_id": str(i["id"] or "null"),
-                                },
-                            }
-                        )
-                    elif status[0].lower() == "error":
-                        output["server_list"].append(
-                            {
-                                "dataTitle": {
-                                    "title_text": "Issue with loadbalancer "
-                                    + str(i["loadbalancer_id"] or "null"),
-                                    "lb_id": str(i["loadbalancer_id"] or "null"),
-                                },
-                                "dataBody": {
-                                    "lb_status": str(ping_result or "null"),
-                                    "amp_status": str(status[1] or "null"),
-                                    "lb_id": str(i["loadbalancer_id"] or "null"),
-                                    "amp_id": str(i["id"] or "null"),
-                                },
-                            }
-                        )
-                    elif ping_result.lower() == "error":
-                        output["server_list"].append(
-                            {
-                                "dataTitle": {
-                                    "title_text": "Issue with amphora "
-                                    + str(i["id"] or "null"),
-                                    "amp_id": str(i["id"] or "null"),
-                                },
-                                "dataBody": {
-                                    "lb_status": str(ping_result or "null"),
-                                    "amp_status": str(status[1] or "null"),
-                                    "lb_id": str(i["loadbalancer_id"] or "null"),
-                                    "amp_id": str(i["id"] or "null"),
-                                },
-                            }
-                        )
-                else:
-                    logging.info("%s is fine.", i["id"])
-
-        else:
-            # Notes problem with accessing api if anything other than 403 or 200 returned
-            logging.critical("We encountered a problem accessing the API")
-            logging.critical("The status code was: %s ", str(amphorae.status_code))
-            logging.critical("The JSON response was: \n %s", str(amph_json))
-            sys.exit(1)
-        return output
-
-    @staticmethod
-    def _check_amphora_status(amphorae):
-        try:
-            amph_json = amphorae.json()
-            return amph_json
-        except requests.exceptions.JSONDecodeError:
-            logging.critical(
-                msg="There was no JSON response \nThe status code was: "
-                + str(amphorae.status_code)
-                + "\nThe body was: \n"
-                + str(amphorae.content)
-            )
-            # pylint: disable=line-too-long
-            return (
-                "There was no JSON response \nThe status code was: "
-                + str(amphorae.status_code)
-                + "\nThe body was: \n"
-                + str(amphorae.content)
-            )
-
-    # pylint: disable=missing-function-docstring
-    @staticmethod
-    def _check_status(amphora):
-        # Extracts the status of the amphora and returns relevant info
-        status = amphora["status"]
-        if status in ("ALLOCATED", "BOOTING", "READY"):
-            return ["ok", status]
-
-        return ["error", status]
-
-    # pylint: disable=invalid-name
-    @staticmethod
-    def _ping_lb(ip):
-        # Runs the ping command to check that loadbalancer is up
-        response = os.system(
-            "ping -c 1 " + ip
-        )  # Might want to update to subprocess as this has been deprecated
-
-        # Checks output of ping command
-        if response == 0:
-            logging.info(msg="Successfully pinged " + ip)
-            return "success"
-
-        logging.info(msg=ip + " is down")
-        return "error"
-
     def check_notify_snapshots(
         self, cloud_account: str, project_id=None, all_projects=False
     ):
+        """
+        Set off and return check for snapshots older than one month
+        """
         # pylint: disable=line-too-long
         output = {
             "title": "Project {p[name]} has an old volume snapshot",
@@ -360,20 +227,22 @@ class CheckActions(Action):
             snapshots = self.check_snapshots(
                 project=project_id, cloud_account=cloud_account
             )
-
-        for snapshot in snapshots:
-            with OpenstackConnection(cloud_name=cloud_account) as conn:
+        with OpenstackConnection(cloud_name=cloud_account) as conn:
+            for snapshot in snapshots:
                 proj = conn.get_project(name_or_id=snapshot["project_id"])
 
-            output["server_list"].append(
-                {"dataTitle": {"name": proj["name"]}, "dataBody": snapshot}
-            )
+                output["server_list"].append(
+                    {"dataTitle": {"name": proj["name"]}, "dataBody": snapshot}
+                )
             # Send email to notify users? projects don't have contact details :/
 
-            return output
+        return output
 
     @staticmethod
     def check_snapshots(cloud_account, project: str):
+        """
+        Check for snapshots that haven't been updated in more than a month.
+        """
         snap_list = []
         with OpenstackConnection(cloud_name=cloud_account) as conn:
             volume_snapshots = conn.list_volume_snapshots(
