+++ sensors/openstack.loadbalancers.yaml
@@ -1,0 +1,3 @@
+class_name: "CheckSensors"
+entry_point: "src/openstack_check_sensors.py"
+description: Check for amphorae with an error status and find out if it is an issue with the amphora or loadbalancer
